mochi_player/components/mpv_player/requirements.txt
pympv==0.6.0
PyQt5==5.14.1